requirements.txt
Faker~=8.14.0
setuptools~=58.1.0
requests~=2.26.0
Pillow~=8.3.2
numpy~=1.21.2
opencv-python~=4.5.3.56
ffpyplayer~=4.3.2
tensorflow~=2.6.0
tensorflow-gpu~=2.6.0